requirements.txt
Django==1.10.6
numpy==1.12.0
olefile==0.44
Pillow==4.0.0
pytz==2016.10
opencv-python==3.2.0